\name{googlebigquery-package}
\alias{googlebigquery-package}
\alias{googlebigquery}
\docType{package}
\title{Google Bigquery R Client}
\description{
R client library for communicating with the Google Bigquery service
}
\details{
\tabular{ll}{
Package: \tab googlebigquery\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-07\cr
License: \tab Apache 2.0\cr
LazyLoad: \tab yes\cr
}
}

\author{Benjamin Yolken <yolken@google.com>}
\references{
See \url{http://code.google.com/apis/bigquery}.
}

\keyword{bigquery}
\keyword{google}
